requirements_dev.txt
-r requirements.txt
mock
freezegun
nose
testing.redis
